server/requirements.txt
fastapi
uvicorn
httpx
aiocache
pytest
pytest-asyncio
python-dotenv